+++ lib/src/formatters/token.dart
@@ -2,7 +2,7 @@
 
 import 'package:moment_dart/src/localizations.dart';
 
-typedef FormatterFn = String Function(DateTime, FormatterToken, Localization);
+typedef FormatterFn = String Function(DateTime, FormatterToken, MomentLocalization);
 typedef FormatterTokenFn = String Function(DateTime);
 
 enum FormatterToken {