+++ lib/src/localizations/all.dart
@@ -1,0 +1,6 @@
+library moment_dart_localizations;
+
+export './en_US.dart';
+export './mn_MN.dart';
+export './mn_TM.dart';
+export './mn_TM_tn.dart';